% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cbs_lookup}
\alias{cbs_lookup}
\title{Look-up table for CBS statistical areas on the Veluwe}
\format{
A data frame with 2 rows and 3 variables:
\describe{
\item{scope}{Character indicating the id of the geographic scope of the Veluwe as used in \link{create_veluwe}}
\item{feature}{Character indicating the name of the feature in CBS dataset at PDOK.}
\item{name}{Character vector indicating the names of administrative areas.}
}
}
\usage{
cbs_lookup
}
\description{
Look-up table to be used in \link{create_veluwe} in case the user selected one of the two administrative scopes of the Veluwe. Table matches the scope ids to the feature names in the CBS dataset and the names of the administrative areas.
}
\seealso{
The script to create the CBS look-up table: data-raw/cbs_lookup.R
}
\keyword{datasets}
