% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-veluwe.R
\name{create_veluwe}
\alias{create_veluwe}
\title{Create geospatial sf objects of the Veluwe}
\usage{
create_veluwe(scope = c("natura2000", "concave", "corop", "quarter", "nphv"))
}
\arguments{
\item{scope}{Character indicating the preferred geospatial scope of the Veluwe. Options: "natura2000", "concave", "corop", "quarter", "nphv". See Details for more info.}
}
\value{
a \link[sf]{sf} object
}
\description{
This function is used to create geospatial \link[sf]{sf} objects of the Veluwe.
}
\details{
This function is used to create a geospatial \link[sf]{sf} object of the Veluwe. There are multiple options:
\itemize{
\item \code{natura2000}: The Veluwe Natura2000 area. Natura 2000 is a network of nature protection areas in the European Union. The Veluwe is the largest Natura 2000 area in the Netherlands on land. It consists of sites that are classified under the Birds Directive and Habitat Directive. More info: \url{https://biodiversity.europa.eu/natura2000/en/natura2000}
\item \code{concave}: A concave hull of the Veluwe Natura2000 area. A \href{https://postgis.net/docs/ST_ConcaveHull.html}{concave hull} is a concave polygon containing all the line segments of the Natura2000 area. Consider it the minimum outer smooth of the Natura2000 area.
\item \code{corop}: The Veluwe COROP region. \href{https://en.wikipedia.org/wiki/COROP}{COROP} is the \href{https://ec.europa.eu/eurostat/web/nuts}{NUTS} level-3 statistical classification of the Netherlands, which is used for statistical and demographic purposes. It comprises the natural areas as well as socio-economic areas that are considered part of the Veluwe.
\item \code{quarter}: The \href{https://en.wikipedia.org/wiki/Veluwe_Quarter}{Veluwe Quarter} was one of four quarters in the Duchy of Guelders. It comprises the Veluwe COROP region plus five municipalities that make up the Veluwezoom (from west to east: Renkum, Arnhem, Rozendaal, Rheden, Brummen).
\item \code{nphv}: De Hoge Veluwe National Park, a privately owned area that is part of the Natura2000 area. More info: \url{https://www.hogeveluwe.nl/en/}
}
}
