% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Get elevation data for the Veluwe}
\usage{
get_elevation(veluwe, topo = c("dtm", "dsm"))
}
\arguments{
\item{veluwe}{\link[sf]{sf} object of the Veluwe. Output from \link{create_veluwe}.}

\item{topo}{Character indicating the topographic model. Options: "dtm", "dsm". See Details for more info.}
}
\value{
a spatial raster
}
\description{
This function is used to retrieve elevation data from \href{https://www.ahn.nl/}{Actueel Hoogtebestand Nederland} (AHN). AHN provides data as a digital terrain model (DTM) or as a digital surface model (DSM).
}
\details{
This function is used to retrieve elevation data from \href{https://www.ahn.nl/}{Actueel Hoogtebestand Nederland} (AHN). AHN provides two products. Both products are a 5 x 5 meter raster, averaged over 50 x 50 cm cells, each of which is assigned a value by interpolating LiDAR aerial survey points using squared inverse distance weighting. More info (in Dutch): \url{https://www.ahn.nl/producten}.
\itemize{
\item \code{dtm}: the digital terrain model (DTM) represents elevation of the elevation of the ground.
\item \code{dsm}: the digital surface model (DSM) represents elevation of the tallest surfaces at a point (e.g., trees, buildings).
}
}
